Move mlir_to_verilog to magma/backend/mlir
